Generics contracts now ignore unsupported methods.
With the addition of method fallbacks, there are a few choices for how generics
contracts might treat unsupported methods.  They might (a) always reject them,
(b) always accept them, or (c) wrap fallbacks if present and otherwise reject
them.  The previous implementation was (a); now it is (b), which wraps
everything implemented for the specific generic type, and trusts the
implementations of the fallbacks, since they can only interact with the
underlying type through the supported operations anyway.  This keeps the
fallbacks from pervading the implementation of generics and also reduces
contract wrapping.